refactor(rocm): move the rope and norm AITER blocks into _rocm/
Same shape as the page/activation move in the parent commit: the AITER
module loader, the backend selectors and the routing bodies move to
flashinfer/_rocm/<op>.py, and each upstream call site keeps only a
delegation.

Footprint against the upstream merge base (2628bebcf):

    norm.py  +131/-3  ->  +48/-3
    rope.py  +144/-7  ->  +97/-7

norm.py evacuates 63%. rope.py only 33%, for the same reason
activation.py did: what remains is a backend= parameter on upstream's
own function -- the docstring entry, the argument validation and the
up-front require_aiter for the explicit opt-in -- which cannot move
without forking the function itself.

rope's two call sites (out-of-place and in-place) differ only in which
tensors receive the output, so they share one maybe_ helper; the
in-place site passes query/key as their own outputs.

Three test files reached into flashinfer.norm / flashinfer.rope for
moved selectors and follow them to flashinfer._rocm.*:
test_rmsnorm_aiter_hip.py, test_norm_hip.py, test_rope_hip.py.
